Add Controller and Repository tests (#2)

+++ src/main/java/com/deepamin/springreactawstodo/controller/ToDoService.java
@@ -1,11 +1,13 @@
 package com.deepamin.springreactawstodo.controller;
 
+import com.deepamin.springreactawstodo.exception.ToDoNotFoundException;
 import com.deepamin.springreactawstodo.model.ToDo;
 import com.deepamin.springreactawstodo.repository.ToDoRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @AllArgsConstructor
@@ -22,11 +24,16 @@ public class ToDoService {
         toDoRepository.save(toDo);
     }
 
-    public void deleteToDoById(Long id) {
+    public void deleteToDoById(Long id) throws ToDoNotFoundException {
+        if (!toDoRepository.existsById(id)) {
+            throw new ToDoNotFoundException("ToDo with id " + id + " doesn't exist.");
+        }
         toDoRepository.deleteById(id);
     }
 
-//    public void editToDo(Long id, ToDo toDo) {
-//       toDoRepository.findById(id).orElseThrow(() -> new IllegalStateException("No To Do item associated for editing with id " + id));
+//    @Transactional
+//    public void editToDo(ToDo toDo) {
+//       toDoRepository.findById(toDo.getId()).orElseThrow(() -> new IllegalStateException("No To Do item associated for editing with id " + toDo.getId()));
+//       toDoRepository.save(toDo);
 //    }
 }
